requirements.txt
django
numpy
scikit-learn
pandas
django-cors-headers
djangorestframework
setuptools
tensorflow